A bunch of changes after moving to the new GIT server

+++ src/tramberend7215/a01/Main.java
@@ -27,7 +27,7 @@ public class Main {
         }
     }
 
-    static Vec3 pixelColor(int x, int y) {
+    static Vec3 pixelColor(double x, double y) {
         return vec3(0.5, 0.5, 0.5);
     }
 }
